Fix calculation inaccurancy when report period is not month
When calculation hourly_rate for ChargebackRateDetail with monthly rate,
we had assumed that lenght of report interval was good indicator of
hours in month. This is however truth only for monthly reports.

For example, assume monthly rate of $4 for fixed compute. On monthly
report we charge correctly $4 (the hourly_rate is ($4 / 1.month). On
weekly report, however we charge $4 per each week, as we assume that
hourly_rate is ($4 / 1.week).

(cherry picked from commit c3d3e8539037ac42bb12f5b4199e7231f0109430)