When building a graph of Dtocollections, put transitive edges to the end
When building a graph of Dtocollections, put transitive edges to the
end, this allow us to process non transitive edges first and deal with
the possible cycles. Then we step to transitive edges, which are more
likely to have a cycle and moving non-transitive edge in such case could
cause non resolvable cycle (not effectively resolvable actually).