Fixing bug where a string value in an expression was incorrectly quoted because it incorrectly identified the string as "model_name-field_name".
We are using only a regular expression to for this validation. However, there are string values that will also match this
pattern that are not model and field names and are instead, plain strings. So, as an additional check, we are now
validating that the model_name portion of the string is an actual class.

This bug was causing the OOTB report "VMware Tools Versions" to blow up because the conditions include:
{"field"=>"Vm.guest_applications-name", "value"=>"VMware"}.

In this case the guest application name is ""NetworkManager-team". This change will not detect this as a model and field.