Correct requester in Dialog#content
`Dialog#content` depends on `ResourceActionWorkflow`, which requires a
requester object. The one that was being passed as the unassigned
variable `@auth_user_obj` (presumably copied/pasted from the API code) -
effectively `nil`. This doesn't seem to matter because the only method
called on the workflow (`#dialog`) does not use the requester
anywhere. Still, it would be better to new up this object with a valid
requester for the sake of fewer surprises.