Validate users, dialogs, and miq_groups properly by region
Previously we were trying to validate the uniqueness of fields by
region using `in_my_region` with `conditions`, but this resulted
in anything not in the current server's region being invalid.

This commit adds a custom validation to properly validate a given
field's uniqueness scoped by the saved row's region.

Fixes #13274

Move the validation to a real validation class rather than a method

This allows us to do nice things like chain the validations together
and test the validation in isolation.