Fix times in chargeback specs
These tests are using timecop to travel to specific
times. Unfortunately, those times are specific to the local timezone,
and are only after represented as UTC via
`Time.parse(time).utc`. Consequently, in some timezones the tests ask
for metrics for the wrong day and return nothing. This updates them to
parse UTC times everywhere so it should be the same for everyone.

Follow up to https://github.com/ManageIQ/manageiq/pull/13282