Automate Retry with Server Affinity
https://bugzilla.redhat.com/show_bug.cgi?id=1408850

A automate method retry can happen in any of the servers in the zone.
If a customers automate method had started a process on a specific
server, they would like to have the retry happen on the same server.
This PR optionally allows an automate method to have the retries happen
on the same server using an attribute in the $evm.root object accessible
to automate methods. The name of the attribute is called
ae_retry_server_affinity

e.g. To enable the retry to happen on the same server use

$evm.root['ae_retry_server_affinity'] = true