Don't let Condition.subst make things taggable
This method changes the class definition of the thing passed to it,
which breaks tagging if a Tag object is passed to it.

For this to work, it was necessary to take the required logic out of
`ArTaggable` and put it at the class level of `Tag`.

Fixes https://github.com/ManageIQ/manageiq/issues/13180