Don't send 'next_run=false' to oVirt 3.6
In oVirt virtual machines have two sets of settings: the settings used
by the current running virtual machine, and the settings that will be
used the next time the virtual machine is started. To select which one
to update it is necessary to use the 'next_run' parameter. For example,
to update the memory that will be used the next time is started the
API request should be like this:

  PUT /ovirt-engine/api/vms/{vm:id};next_run=true

The value 'false' should tell the server to update the current running
configuration. But in version 3.6 of oVirt it is just ignored, the mere
presence of the 'next_run' parameter is interpreted as a request to
modify the next start settings, regardless of its value.

Currently the oVirt provider uses 'next_run=false' explicitly, and as a
result it fails to update the current settings. This patch changes the
provider so that it doesn't send the 'next_run' parameter when its value
is 'false'.

https://bugzilla.redhat.com/1356468