The default_scope should not rely on cols that are soon to be removed
This one is problematic. We don't like default scopes and we don't like
joins in default scopes.

However, it is interconnected. The rates editor ui depends on this
default scope and the editor is very complex. By detaching chargeable
field from chargeback_rate_detail we are making room for rate editor to
become less comlex. Thus, enabling rate editor not to depend on this
default scope, thus, enabling removal from this default scope. It's
journey.