DSL
A `Persister` can `include ManagerRefresh::Inventory::AutomationManager`
to have access to all inventory collections an AutomationManager usually
persists. Then we call the DSL `has_inventory` to define a method on the
`Persister` that returns the InventoryCollection

Documentation

`ManagerRefresh::Inventory::Core` will have all CIs that are available
for saving. For all **Managers** there is
`ManagerRefresh::Inventory::<Type>Manager` which has all the CIs that
are available for that Manager.