HACK: Backport rails/rails#26121
This was added in Rails v5.0.1:

https://github.com/rails/rails/pull/26121

And fixes a somewhat esoteric bug with counts in rails by removing an
arel ast lookup.

This was backported here by prepending the affected methods after that
patch as an anonymous prior to the that then overwrites some of those
same methods.  This was the simplest way of updating the methods and
retaining their original state before we then update them (instead of
doing it in one fell swoop)  This means there is a bit of extra code,
but it is easier to diff the hack against the code it was pulled from:

https://github.com/rails/rails/blob/38cfa6a/activerecord/lib/active_record/relation/calculations.rb