Add embedded_ansible to the list of roles that need apache
We handle apache in a weird way.

Rather than starting it when we need it, we always issue a restart
when the server comes up (in MiqServer::EnvironmentManagement.prep_apache_proxying)
but then shut it down if none of our active roles require it.

That definition of requirement is governed by this array of roles
so just adding the role name to the array will keep apache up if the
role is active.