Fix PasswordMixin.encrypted_columns
This mixin uses a class variable which has the (perhaps surprising)
behavior of sharing data up and down the inheritance
tree. Consequently, this leads to quirky behavior such as:

```
irb(main):001:0> Authentication.encrypted_columns
=> ["auth_key", "auth_key_encrypted", "password", "password_encrypted", "service_account", "service_account_encrypted"]
irb(main):002:0> MiqDatabase.encrypted_columns
=> ["auth_key", "auth_key_encrypted", "password", "password_encrypted", "service_account", "service_account_encrypted", "csrf_secret_token", "csrf_secret_token_encrypted", "session_secret_token", "session_secret_token_encrypted"]
irb(main):003:0> Authentication.encrypted_columns
=> ["auth_key", "auth_key_encrypted", "password", "password_encrypted", "service_account", "service_account_encrypted", "csrf_secret_token", "csrf_secret_token_encrypted", "session_secret_token", "session_secret_token_encrypted"]
```

You will notice here that the first invocation of
`Authentication.encrypted_columns` appears correct in returning the
encrypted columns as declared in that
model. `MiqDatabase.encrypted_columns` returns all its columns *plus*
those of `Authentication`. The second invokation of
`Authentication.encrypted_columns` now has `MiqDatabase`'s
concatenated to it.

The solution is to make this a `class_attribute` that's defined via
`ActiveSupport::Concern`'s `included`. This way, a base class such as
`Authentication` that includes this module can share that data with
its descendants, but that data isn't shared with other hosts that are
not kinds of `Authentication`