Make worker_monitor_drb act like a reader again!
https://bugzilla.redhat.com/show_bug.cgi?id=1438935

For testing in e21d1b9ca, @worker_monitor_drb was changed to use the
method.  Unfortunately, this method isn't a reader, so it would create a
new drb client each time it was called.

It doesn't make sense for this method to make a new drb client each
time, so we can cache this client on first access so it functions like a
attr_reader now.  We can then remove some memoization/caching that
occurs in heartbeat since the getter method handles that for us.