quiesce_time_allowance per worker is not needed
We already check quiesce_loop_timeout and kill all workers if the loop
timeout is reached.  It doesn't matter what quiesce_time_allowance is
for a worker if the quiesce_loop_timeout is lower, since the loop
timeout will expire first and cause all workers to be killed.

It doesn't seem necessary to have two timeouts and doing it at a worker
level seems too complex.

The original commit and bugzids don't indicate this per worker class
timeout as a feature so it must have been added as something we might
use in the future.

Below is the original commit found by NickL :tada:

commit aa2f40fa4504bae0854e6ac194b89a43845e9934
Author: Joe Rafaniello <jrafanie@redhat.com>
Date:   Tue Mar 29 23:21:47 2011 +0000

    Add a synchronous stop of the server and call it from EvmApplication(used by rake evm:stop/restart and black console)
    Add quiesce logic to allow queue worker types to exit gracefully, other types are killed, and only continue shutdown of the server when they're all stopped or killed.
    If provided, use :quiesce_loop_timeout in worker_monitor section of configuration
    If provided, use :quiesce_time_allowance in each worker's settings
    bugzid:11862,11825,11760,11719