Skip test setup in prod env
In bin/setup and bin/update, the task of test:vmdb:setup is run even in
production, but will be disabled because we only create that rake task
when RSpec is present (which it is not in the production group).

Disabling calling that task when RAILS_ENV=production is set avoids an
error that can be confusing at first glance.